Fixed all compilation issues on VS 2022

+++ src/server/game/Instances/InstanceSaveMgr.cpp
@@ -425,7 +425,7 @@ void InstanceSaveManager::LoadResetTimes()
                 continue;
 
             // the reset_delay must be at least one day
-            uint32 period = uint32(((mapDiff->GetRaidDuration() * sWorld->getRate(RATE_INSTANCE_RESET_TIME)) / DAY) * DAY);
+            uint32 period = static_cast<uint32>(((mapDiff->GetRaidDuration() * static_cast<float>(sWorld->getRate(RATE_INSTANCE_RESET_TIME))) / static_cast<float>(DAY)) * static_cast<float>(DAY));
             if (period < DAY)
                 period = DAY;
 
@@ -473,7 +473,7 @@ time_t InstanceSaveManager::GetSubsequentResetTime(uint32 mapid, Difficulty diff
     }
 
     time_t diff = sWorld->getIntConfig(CONFIG_INSTANCE_RESET_TIME_HOUR) * HOUR;
-    time_t period = uint32(((mapDiff->GetRaidDuration() * sWorld->getRate(RATE_INSTANCE_RESET_TIME)) / DAY) * DAY);
+    time_t period = static_cast<uint32>(((mapDiff->GetRaidDuration() * static_cast<float>(sWorld->getRate(RATE_INSTANCE_RESET_TIME))) / static_cast<float>(DAY)) * static_cast<float>(DAY));
     if (period < DAY)
         period = DAY;
 
@@ -755,3 +755,4 @@ uint32 InstanceSaveManager::GetNumBoundGroupsTotal() const
 
     return ret;
 }
+
